journal: move rados_ctx_callback to utils namespace
