+++ src/journal/Utils.h
@@ -5,6 +5,7 @@
 #define CEPH_JOURNAL_UTILS_H
 
 #include "include/int_types.h"
+#include "include/rados/librados.hpp"
 #include <string>
 
 namespace journal {
@@ -13,6 +14,8 @@ namespace utils {
 std::string get_object_name(const std::string &prefix, uint64_t number);
 
 std::string unique_lock_name(const std::string &name, void *address);
+
+void rados_ctx_callback(rados_completion_t c, void *arg);
 
 } // namespace utils
 } // namespace journal
